Create README.md
This project aims to develop a Support Vector Machine (SVM) model for predicting breast cancer. The goal is to classify whether a breast tumor is benign or malignant based on various features extracted from breast cancer cell images. The project uses the Wisconsin Breast Cancer Dataset, a well-known dataset in the machine learning community, to train and evaluate the SVM model